+++ lib/scalingo/bearer_token.rb
@@ -1,12 +1,13 @@
+require "jwt"
+
 module Scalingo
   class BearerToken
     attr_reader :expires_at
-    attr_writer :raise_on_expired
 
-    def initialize(value, expires_at: nil, raise_on_expired: false)
+    def initialize(value)
       @value = value
-      @expires_at = expires_at if expires_at
-      @raise_on_expired = raise_on_expired
+
+      read_expiration!
     end
 
     # :nocov:
@@ -26,18 +27,22 @@ module Scalingo
     end
     # :nocov:
 
-    def raise_on_expired?
-      @raise_on_expired
-    end
-
     def expired?
       expires_at && expires_at <= Time.now
     end
 
     def value
-      raise Error::ExpiredToken if expired? && raise_on_expired?
+      raise Error::ExpiredToken if expired?
 
       @value
+    end
+
+    private
+
+    def read_expiration!
+      payload, _ = JWT.decode(@value, nil, false)
+
+      @expires_at = Time.at(payload["exp"]) if payload["exp"]
     end
   end
 end