+++ lib/scalingo/configuration.rb
@@ -1,36 +1,15 @@
 require "active_support"
 require "active_support/core_ext/numeric/time"
 require "scalingo/version"
-require "ostruct"
 
 module Scalingo
   class Configuration
     ATTRIBUTES = [
-      # URL to the authentication API
-      # :auth,
-
-      # URL to the billing API
-      # :billing,
-
-      # List of regions under the form {"region_id": root_url}
-      # :regions,
-
       # Default region. Must match a key in `regions`
       :default_region,
 
       # Configure the User Agent header
       :user_agent,
-
-      # For how long is a bearer token considered valid (it will raise passed this delay).
-      # Set to nil to never raise.
-      :exchanged_token_validity,
-
-      # Raise an exception when trying to use an authenticated connection without a bearer token set
-      # Having this setting to true prevents performing requests that would fail due to lack of authentication headers.
-      :raise_on_missing_authentication,
-
-      # Raise an exception when the bearer token in use is supposed to be invalid
-      :raise_on_expired_token,
 
       # These headers will be added to every request. Individual methods may override them.
       # This should be a hash or a callable object that returns a hash.
@@ -46,9 +25,6 @@ module Scalingo
       new(
         default_region: :osc_fr1,
         user_agent: "Scalingo Ruby Client v#{Scalingo::VERSION}",
-        exchanged_token_validity: 1.hour,
-        raise_on_missing_authentication: true,
-        raise_on_expired_token: false,
         additional_headers: {}
       )
     end