+++ spec/scalingo/configuration_spec.rb
@@ -24,15 +24,16 @@ RSpec.describe Scalingo::Configuration do
   end
 
   describe "faraday adapter" do
-    let(:scalingo) { Scalingo::Client.new(config).tap { |s| s.authenticate_with(bearer_token: "some-token") } }
+    let(:jwt) { Scalingo.generate_test_jwt(duration: 1.hour) }
+    let(:scalingo) { Scalingo::Client.new(config).tap { |s| s.authenticate_with(bearer_token: jwt) } }
     let(:client) { Scalingo::API::Client.new("http://example.test", scalingo: scalingo) }
 
     context "when unspecified" do
       let(:config) { {} }
 
       it "uses the default one when unspecificied" do
-        expect(client.authenticated_connection.adapter).to eq Faraday::Adapter::NetHttp
-        expect(client.unauthenticated_connection.adapter).to eq Faraday::Adapter::NetHttp
+        expect(client.connection.adapter).to eq Faraday::Adapter::NetHttp
+        expect(client.guest_connection.adapter).to eq Faraday::Adapter::NetHttp
       end
     end
 
@@ -40,8 +41,8 @@ RSpec.describe Scalingo::Configuration do
       let(:config) { {faraday_adapter: :yo} }
 
       it "uses the default one when unspecificied" do
-        expect { client.authenticated_connection.adapter }.to raise_error(Faraday::Error)
-        expect { client.unauthenticated_connection.adapter }.to raise_error(Faraday::Error)
+        expect { client.connection.adapter }.to raise_error(Faraday::Error)
+        expect { client.guest_connection.adapter }.to raise_error(Faraday::Error)
       end
     end
 
@@ -49,8 +50,8 @@ RSpec.describe Scalingo::Configuration do
       let(:config) { {faraday_adapter: :test} }
 
       it "uses the default one when unspecificied" do
-        expect(client.authenticated_connection.adapter).to eq Faraday::Adapter::Test
-        expect(client.unauthenticated_connection.adapter).to eq Faraday::Adapter::Test
+        expect(client.connection.adapter).to eq Faraday::Adapter::Test
+        expect(client.guest_connection.adapter).to eq Faraday::Adapter::Test
       end
     end
   end